backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Quant Service API"
authors = [{ name = "Dan Kim", email = "daniel@mysingle.io" }]
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "mysingle-quant>=0.1.33",
    "uvicorn[standard]>=0.24.0",
    "pydantic-extra-types>=2.2.0",
    "motor>=3.3.0",
    "pymongo>=4.5.0",
    "aiohttp>=3.9.4",
    "pandas>=2.0.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "vectorbt>=0.25.2",
    "ta-lib>=0.4.0",
    "psutil>=5.9.0",
    "duckdb>=0.6.0",
    "pyjwt>=2.10.1",
    "emails>=0.6",
    "pwdlib[argon2,bcrypt]>=0.1.7",
    "httpx-auth>=0.16.1",
    "fastapi[standard]>=0.117.1",
    "tenacity>=9.1.2",
    "structlog>=25.4.0",
    "lightgbm>=4.6.0",
    "optuna>=4.5.0",
    "scikit-learn>=1.7.2",
    "openai>=2.3.0",
    "mlflow>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.14.0",
    "black>=23.9.0",
    "ruff>=0.7.1",
    "mypy>=1.17.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
ignore = ["F403"] # wildcard imports (schemas/__init__.py에서 사용)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "e2e: End-to-end integration tests (deselect with '-m \"not e2e\"')",
    "slow: Slow-running tests (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
